+++ src/BamRecord.cpp
@@ -634,8 +634,8 @@ void BamRecord::ClipFields(const size_t clipFrom,
     const bool isForwardStrand = (AlignedStrand() == Strand::FORWARD);
 
     // clip seq, quals
-    string sequence = std::move(internal::Clip(Sequence(Orientation::NATIVE), clipFrom, clipLength));
-    QualityValues qualities = std::move(internal::Clip(Qualities(Orientation::NATIVE), clipFrom, clipLength));
+    string sequence = internal::Clip(Sequence(Orientation::NATIVE), clipFrom, clipLength);
+    QualityValues qualities = internal::Clip(Qualities(Orientation::NATIVE), clipFrom, clipLength);
     if (!isForwardStrand) {
         internal::ReverseComplement(sequence);
         internal::Reverse(qualities);
@@ -707,7 +707,7 @@ BamRecord& BamRecord::ClipToQuery(const Position start,
     if (IsMapped()) {
 
         // fetch a 'working copy' of CIGAR data
-        Cigar cigar = std::move(impl_.CigarData());
+        Cigar cigar = impl_.CigarData();
 
         // clip leading CIGAR ops
         size_t referencePositionOffset = 0;
@@ -820,7 +820,7 @@ BamRecord& BamRecord::ClipToReferenceForward(const PacBio::BAM::Position start,
     const Position newTEnd   = std::min(origTEnd, end);
 
     // fetch a 'working copy' of CIGAR data
-    Cigar cigar = std::move(impl_.CigarData());
+    Cigar cigar = impl_.CigarData();
 
     // we're going to skip query sequence outside aligned region
     size_t queryPosRemovedFront = 0;
@@ -950,7 +950,7 @@ BamRecord& BamRecord::ClipToReferenceReverse(const PacBio::BAM::Position start,
     const Position newTStart = std::max(origTStart, start);
     const Position newTEnd   = std::min(origTEnd, end);
 
-    Cigar cigar = std::move(impl_.CigarData());
+    Cigar cigar = impl_.CigarData();
 
     size_t queryPosRemovedFront = 0;
     size_t queryPosRemovedBack  = 0;
@@ -1119,10 +1119,10 @@ string BamRecord::FetchBases(const string& tagName,
     string bases;
     Orientation current;
     if (isBamSeq) { // SEQ stored in genomic orientation
-        bases = std::move(impl_.Sequence());
+        bases = impl_.Sequence();
         current = Orientation::GENOMIC;
     } else { // all tags stored in native orientation
-        bases = std::move(FetchBasesRaw(tagName));
+        bases = FetchBasesRaw(tagName);
         current = Orientation::NATIVE;
     }
 
@@ -1162,15 +1162,14 @@ Frames BamRecord::FetchFramesRaw(const string& tagName) const
 
     // lossy frame codes
     if (frameTag.IsUInt8Array()) {
-        const vector<uint8_t> codes = std::move(frameTag.ToUInt8Array());
-        frames = std::move(Frames::Decode(codes));
+        const vector<uint8_t> codes = frameTag.ToUInt8Array();
+        frames = Frames::Decode(codes);
     }
 
     // lossless frame data
     else {
         assert(frameTag.IsUInt16Array());
-        const vector<uint16_t> losslessFrames = std::move(frameTag.ToUInt16Array());
-        frames.Data(std::move(losslessFrames));
+        frames.Data(frameTag.ToUInt16Array());
     }
 
     return frames;
@@ -1222,7 +1221,7 @@ vector<float> BamRecord::FetchPhotons(const string& tagName,
         return vector<float>();
     if(!frameTag.IsUInt16Array())
         throw std::runtime_error("Photons are not a uint16_t array, tag " + tagName);
-    vector<uint16_t> data = std::move(frameTag.ToUInt16Array());
+    vector<uint16_t> data = frameTag.ToUInt16Array();
 
     // put in requested orientation
     internal::OrientTagDataAsRequested(&data,
@@ -1260,10 +1259,10 @@ QualityValues BamRecord::FetchQualities(const string& tagName,
     QualityValues quals;
     Orientation current;
     if (isBamQual) { // QUAL stored in genomic orientation
-        quals = std::move(impl_.Qualities());
+        quals = impl_.Qualities();
         current = Orientation::GENOMIC;
     } else {        // all tags stored in native orientation
-        quals = std::move(FetchQualitiesRaw(tagName));
+        quals = FetchQualitiesRaw(tagName);
         current = Orientation::NATIVE;
     }
 
@@ -1525,7 +1524,7 @@ Frames BamRecord::IPDRaw(Orientation orientation) const
 
     // lossy frame codes
     if (frameTag.IsUInt8Array()) {
-        const vector<uint8_t> codes = std::move(frameTag.ToUInt8Array());
+        const vector<uint8_t> codes = frameTag.ToUInt8Array();
         const vector<uint16_t> codes16(codes.begin(), codes.end());
         frames.Data(std::move(codes16));
     }
@@ -1533,8 +1532,7 @@ Frames BamRecord::IPDRaw(Orientation orientation) const
     // lossless frame data
     else {
         assert(frameTag.IsUInt16Array());
-        const vector<uint16_t> losslessFrames = std::move(frameTag.ToUInt16Array());
-        frames.Data(std::move(losslessFrames));
+        frames.Data(frameTag.ToUInt16Array());
     }
 
     // return in requested orientation
@@ -1556,7 +1554,7 @@ Frames BamRecord::PulseWidthRaw(Orientation orientation) const
 
     // lossy frame codes
     if (frameTag.IsUInt8Array()) {
-        const vector<uint8_t> codes = std::move(frameTag.ToUInt8Array());
+        const vector<uint8_t> codes = frameTag.ToUInt8Array();
         const vector<uint16_t> codes16(codes.begin(), codes.end());
         frames.Data(std::move(codes16));
     }
@@ -1564,8 +1562,7 @@ Frames BamRecord::PulseWidthRaw(Orientation orientation) const
     // lossless frame data
     else {
         assert(frameTag.IsUInt16Array());
-        const vector<uint16_t> losslessFrames = std::move(frameTag.ToUInt16Array());
-        frames.Data(std::move(losslessFrames));
+        frames.Data(frameTag.ToUInt16Array());
     }
 
     // return in requested orientation